+++ src/components/resultPage/exportOptions/ExportOptions.jsx
@@ -1,11 +1,11 @@
 import './ExportOptions.css';
 import { Card, Col, Container, Image, Offcanvas, Row } from 'react-bootstrap';
 import Mail from '../../../api/mail/Mail';
-import mail from '../../../../public/media/envelope-at.svg';
-import upload from '../../../../public/media/upload.svg';
+import mail from '../../../media/envelope-at.svg';
+import upload from '../../../media/upload.svg';
 import StoreState from '../../../api/StoreState';
 import SaveRoomPlan from '../../../api/SaveRoomPlan';
-import download from '../../../../public/media/download.svg';
+import download from '../../../media/download.svg';
 import RevagerLiteExport from '../../../api/RevagerLiteExport';
 import React from 'react';
 import { useRoomSlots } from '../../shared/context/RoomSlotContext';
@@ -97,18 +97,30 @@ function ExportOptions ({ onHide, ...props }) {
                             <Col>
                                 <Card style={{ backgroundColor: '#cccccc' }}>
                                     <Card.Body>
-                                        <Card.Title>Room Plan</Card.Title>
+                                        <Card.Title>Room Plan PDF</Card.Title>
                                         <CustomButton
                                             backgroundColor={'#B0D7AF'}
-                                            onButtonClick={() => new SaveRoomPlan(roomSlots).runSave()}
-                                            toolTip={'Download results in pdf format'}
+                                            onButtonClick={() => new SaveRoomPlan(roomSlots).runSaveAsTable()}
+                                            toolTip={'Download Room Plan as simple Table'}
                                         >
                                             <Image
                                                 src={download}
                                                 alt="exportResult"
                                                 className={'button-icon'}
                                             />
-                                            Download pdf
+                                            Download Table
+                                        </CustomButton>
+                                        <CustomButton
+                                            backgroundColor={'#B0D7AF'}
+                                            onButtonClick={() => new SaveRoomPlan(roomSlots).runSaveAsMatrix()}
+                                            toolTip={'Download Room Plan as Matrix'}
+                                        >
+                                            <Image
+                                                src={download}
+                                                alt="exportResult"
+                                                className={'button-icon'}
+                                            />
+                                            Download Matrix
                                         </CustomButton>
                                     </Card.Body>
                                 </Card>